# docs/api/whatsapp-ai-openapi.yaml
openapi: 3.0.3
info:
  title: autolumiku WhatsApp AI API
  description: API for WhatsApp AI Automation - customer chat & staff commands management
  version: 1.0.0
  contact:
    name: autolumiku Support
    email: support@autolumiku.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.autolumiku.com/v1
    description: Production server
  - url: https://staging-api.autolumiku.com/v1
    description: Staging server
  - url: http://localhost:3000/api/v1
    description: Development server
security:
  - BearerAuth: []
paths:
...
  /whatsapp-ai/analytics:
    get:
      tags:
        - Analytics
      summary: Get comprehensive analytics
      description: Retrieve analytics with time series, intent breakdown, and staff activity
      operationId: getAnalytics
      parameters:
        - name: tenantId
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: range
          in: query
          required: false
          schema:
            type: string
            enum: [today, week, month]
            default: week
          description: Time range for analytics
      responses:
        '200':
          description: Analytics retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Analytics'
...
  /webhooks/aimeow:
    post:
      tags:
        - Webhooks
      summary: Aimeow webhook endpoint
      description: Receives WhatsApp events from Aimeow (messages, status updates, connection events)
      operationId: aimeowWebhook
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookPayload'
      responses:
        '200':
          description: Webhook processed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        '400':
          description: Bad request
        '404':
          description: Account not found

    get:
      tags:
        - Webhooks
      summary: Webhook verification
      description: Verify webhook endpoint (used by Aimeow for initial handshake)
      operationId: verifyWebhook
      security: []
      parameters:
        - name: hub.mode
          in: query
          schema:
            type: string
        - name: hub.verify_token
          in: query
          schema:
            type: string
        - name: hub.challenge
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Verification successful
          content:
            text/plain:
              schema:
                type: string
        '403':
          description: Verification failed
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Analytics:
      type: object
      properties:
        overview:
          type: object
          properties:
            totalConversations:
              type: integer
            activeConversations:
              type: integer
            totalMessages:
              type: integer
            aiResponseRate:
              type: integer
            avgResponseTime:
              type: number
            escalationRate:
              type: integer
        performance:
          type: object
          properties:
            aiAccuracy:
              type: integer
            customerSatisfaction:
              type: integer
            resolutionRate:
              type: integer
            firstResponseTime:
              type: number
        timeSeriesData:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
                format: date
              conversations:
                type: integer
              messages:
                type: integer
              escalations:
                type: integer
        intentBreakdown:
          type: array
          items:
            type: object
            properties:
              intent:
                type: string
              count:
                type: integer
              percentage:
                type: integer
        staffActivity:
          type: array
          items:
            type: object
            properties:
              staffPhone:
                type: string
              commandCount:
                type: integer
              successRate:
                type: integer
              lastActive:
                type: string
                format: date-time
...
    WebhookPayload:
      type: object
      required:
        - clientId
        - event
      properties:
        clientId:
          type: string
        event:
          type: string
          enum: [message, status, qr, connected, disconnected]
        data:
          type: object
          description: Event-specific data
...